review create,delete,

+++ src/review/review.controller.ts
@@ -1,14 +1,18 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards,Request, UnauthorizedException } from '@nestjs/common';
 import { ReviewService } from './review.service';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { Review, User } from '@prisma/client';
+import { ESLint } from 'eslint';
 
 @Controller('review')
 export class ReviewController {
   constructor(private readonly reviewService: ReviewService) {}
 
   @Post()
-  create(@Body() createReviewDto: CreateReviewDto) {
+  @UseGuards(AuthGuard('bearer'))
+  create(@Body() createReviewDto: CreateReviewDto,@Request() req) {
     return this.reviewService.create(createReviewDto);
   }
 
@@ -17,8 +21,11 @@ export class ReviewController {
     return this.reviewService.findAll();
   }
 
+  
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  @UseGuards(AuthGuard('bearer'))
+  findOne(@Param('id') id: string,@Request() req) {
+    const user:User=req.user;
     return this.reviewService.findOne(+id);
   }
 
@@ -28,7 +35,15 @@ export class ReviewController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.reviewService.remove(+id);
+  @UseGuards(AuthGuard('bearer'))
+  remove(@Param('id') id: string,@Request() req) {
+    const user:User= req.user;
+    const review:Review= req.review;
+    if(user.id != review.userId){
+      throw new UnauthorizedException("Törlés jelentkezzen be!");
+    }
+    else{
+      return this.reviewService.remove(+id);
+    }
   }
 }
